refactor(git): split command boundary owner
