+++ ci/proof.toml
@@ -698,6 +698,7 @@ coverage = true
 name = "git_subprocess_boundary"
 kind = "rust"
 paths = [
+  "crates/tokmd-git/src/command.rs",
   "crates/tokmd-git/src/lib.rs",
   "crates/tokmd-scan/src/walk/git.rs",
   "crates/tokmd-scan/src/walk/mod.rs",
