+++ crates/tokmd-git/src/lib.rs
@@ -17,42 +17,14 @@
 
 use std::io::{BufRead, BufReader};
 use std::path::{Path, PathBuf};
-use std::process::{Command, Stdio};
+use std::process::Stdio;
 
 use anyhow::{Context, Result};
 pub use tokmd_types::CommitIntentKind;
 
-const GIT_REPO_SHAPING_ENV: &[&str] = &[
-    // Repository and object-store overrides.
-    "GIT_DIR",
-    "GIT_WORK_TREE",
-    "GIT_INDEX_FILE",
-    "GIT_OBJECT_DIRECTORY",
-    "GIT_ALTERNATE_OBJECT_DIRECTORIES",
-    "GIT_COMMON_DIR",
-    "GIT_CEILING_DIRECTORIES",
-    // Git hooks that can execute helper programs from ambient environment.
-    "GIT_SSH",
-    "GIT_SSH_COMMAND",
-    "GIT_ASKPASS",
-    "GIT_PAGER",
-    "GIT_EDITOR",
-    "GIT_PROXY_COMMAND",
-    "GIT_EXTERNAL_DIFF",
-];
+mod command;
 
-/// Create a `Command` for git with process-environment isolation.
-///
-/// Strips repo, worktree, index, object-store, discovery, and execution helper
-/// overrides so inherited environment variables cannot bypass the explicit
-/// `-C` path used by functions in this crate or launch ambient helper programs.
-pub fn git_cmd() -> Command {
-    let mut cmd = Command::new("git");
-    for name in GIT_REPO_SHAPING_ENV {
-        cmd.env_remove(name);
-    }
-    cmd
-}
+pub use command::git_cmd;
 
 #[derive(Debug, Clone)]
 pub struct GitCommit {
@@ -456,49 +428,12 @@ fn contains_word(haystack: &str, word: &str) -> bool {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::collections::BTreeSet;
+    use std::process::Command;
 
     fn test_git(dir: &Path) -> Command {
         let mut cmd = git_cmd();
         cmd.arg("-C").arg(dir);
         cmd
-    }
-
-    #[test]
-    fn git_cmd_removes_repo_shaping_env_overrides() {
-        let removed: BTreeSet<_> = git_cmd()
-            .get_envs()
-            .filter(|(_, value)| value.is_none())
-            .map(|(name, _)| name.to_string_lossy().into_owned())
-            .collect();
-
-        for name in GIT_REPO_SHAPING_ENV {
-            assert!(removed.contains(*name), "missing env_remove for {name}");
-        }
-    }
-
-    #[test]
-    fn git_cmd_removes_execution_helper_env_overrides() {
-        let removed: BTreeSet<_> = git_cmd()
-            .get_envs()
-            .filter(|(_, value)| value.is_none())
-            .map(|(name, _)| name.to_string_lossy().into_owned())
-            .collect();
-
-        for name in [
-            "GIT_SSH",
-            "GIT_SSH_COMMAND",
-            "GIT_ASKPASS",
-            "GIT_PAGER",
-            "GIT_EDITOR",
-            "GIT_PROXY_COMMAND",
-            "GIT_EXTERNAL_DIFF",
-        ] {
-            assert!(
-                removed.contains(name),
-                "missing execution env_remove for {name}"
-            );
-        }
     }
 
     #[test]
